spring-mvc

+++ src/main/java/com/lance/persistence/model/User.java
@@ -1,17 +1,17 @@
 package com.lance.persistence.model;
 
 public class User {
-    private Short id;
+    private int id;
 
     private String name;
 
     private String password;
 
-    public Short getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Short id) {
+    public void setId(int id) {
         this.id = id;
     }
